fix includes, move headers to include file
