+++ ADOL-C/include/adolc/storemanager.h
