+++ ADOL-C/src/adouble.cpp
@@ -21,11 +21,11 @@
 
 ----------------------------------------------------------------------------*/
 
-#include "dvlparms.h"
-#include "oplate.h"
-#include "taping_p.h"
 #include <adolc/ad_types.h>
 #include <adolc/adolc.h>
+#include <adolc/dvlparms.h>
+#include <adolc/oplate.h>
+#include <adolc/taping_p.h>
 #include <cassert>
 
 adouble::adouble() {
