+++ ADOL-C/src/adouble_tl_hov.cpp
@@ -21,8 +21,8 @@
 
 ----------------------------------------------------------------------------*/
 
-#include "dvlparms.h"
 #include <adolc/adtl_hov.h>
+#include <adolc/dvlparms.h>
 #include <cmath>
 #include <iostream>
 #include <limits>
