+++ ADOL-C/src/ampisupportAdolc.cpp
@@ -12,8 +12,8 @@
 ----------------------------------------------------------------------------*/
 
 #include "adolc/ad_types.h"
-#include "oplate.h"
-#include "taping_p.h"
+#include <adolc/oplate.h>
+#include <adolc/taping_p.h>
 #include <cassert>
 #include <cstring>
