+++ ADOL-C/src/drivers/psdrivers.cpp
@@ -13,11 +13,11 @@
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#include "dvlparms.h"
-#include "taping_p.h"
 #include <adolc/adalloc.h>
 #include <adolc/drivers/psdrivers.h>
+#include <adolc/dvlparms.h>
 #include <adolc/interfaces.h>
+#include <adolc/taping_p.h>
 
 #include <math.h>
