+++ ADOL-C/src/drivers/taylor.cpp
@@ -12,10 +12,10 @@
  recipient's acceptance of the terms of the accompanying license file.
 
 ----------------------------------------------------------------------------*/
-#include "taping_p.h"
 #include <adolc/adalloc.h>
 #include <adolc/drivers/taylor.h>
 #include <adolc/interfaces.h>
+#include <adolc/taping_p.h>
 
 #include <math.h>
