+++ ADOL-C/src/fo_rev.cpp
@@ -184,19 +184,19 @@ results   Taylor-Jacobians       ------------          Taylor Jacobians
 
 /****************************************************************************/
 /*                                                       NECESSARY INCLUDES */
-#include "dvlparms.h"
-#include "externfcts_p.h"
-#include "oplate.h"
-#include "taping_p.h"
 #include <adolc/adalloc.h>
+#include <adolc/dvlparms.h>
 #include <adolc/externfcts.h>
+#include <adolc/externfcts_p.h>
 #include <adolc/interfaces.h>
+#include <adolc/oplate.h>
+#include <adolc/taping_p.h>
 
 #include <math.h>
 #include <string.h>
 
 #ifdef ADOLC_MEDIPACK_SUPPORT
-#include "medipacksupport_p.h"
+#include <adolc/medipacksupport_p.h>
 #endif
 #ifdef ADOLC_AMPI_SUPPORT
 #include "ampi/ampi.h"
