+++ ADOL-C/src/interfaces.cpp
@@ -13,8 +13,8 @@
 
 ----------------------------------------------------------------------------*/
 
-#include "dvlparms.h"
 #include <adolc/adalloc.h>
+#include <adolc/dvlparms.h>
 #include <adolc/interfaces.h>
 
 /****************************************************************************/
