+++ ADOL-C/src/medipacksupport.cpp
@@ -12,12 +12,12 @@
 ----------------------------------------------------------------------------*/
 
 #include "adolc/ad_types.h"
-#include "oplate.h"
-#include "taping_p.h"
+#include <adolc/oplate.h>
+#include <adolc/taping_p.h>
 
 #ifdef ADOLC_MEDIPACK_SUPPORT
 
-#include "medipacksupport_p.h"
+#include <adolc/medipacksupport_p.h>
 #include <vector>
 
 struct AdolcMeDiAdjointInterface : public medi::AdjointInterface {
