+++ ADOL-C/src/param.cpp
@@ -12,10 +12,10 @@
 
 ----------------------------------------------------------------------------*/
 
-#include "dvlparms.h"
-#include "oplate.h"
-#include "taping_p.h"
 #include <adolc/ad_types.h>
+#include <adolc/dvlparms.h>
+#include <adolc/oplate.h>
+#include <adolc/taping_p.h>
 
 pdouble::pdouble(const double pval) {
 #include <limits>
