+++ ADOL-C/src/rpl_malloc.cpp
@@ -13,7 +13,7 @@
 
 ----------------------------------------------------------------------------*/
 
-#include "rpl_malloc.h"
+#include <adolc/rpl_malloc.h>
 
 #undef ADOLC_NO_MALLOC
 #undef ADOLC_NO_REALLOC
