+++ ADOL-C/src/storemanager.cpp
@@ -12,11 +12,11 @@
  recipient's acceptance of the terms of the accompanying license file.
 
 ---------------------------------------------------------------------------*/
-#include "checkpointing_p.h"
-#include "dvlparms.h"
-#include "taping_p.h"
 #include <adolc/adalloc.h>
+#include <adolc/checkpointing_p.h>
+#include <adolc/dvlparms.h>
 #include <adolc/revolve.h>
+#include <adolc/taping_p.h>
 
 #include <cassert>
 #include <cstring> // For memset
