+++ ADOL-C/src/tape_handling.cpp
@@ -12,14 +12,14 @@
  recipient's acceptance of the terms of the accompanying license file.
 
 ---------------------------------------------------------------------------*/
-#include "checkpointing_p.h"
-#include "dvlparms.h"
-#include "taping_p.h"
 #include <adolc/adalloc.h>
+#include <adolc/checkpointing_p.h>
+#include <adolc/dvlparms.h>
 #include <adolc/revolve.h>
+#include <adolc/taping_p.h>
 
 #ifdef ADOLC_MEDIPACK_SUPPORT
-#include "medipacksupport_p.h"
+#include <adolc/medipacksupport_p.h>
 #endif
 
 #include <algorithm>
